+++ R/finalfit_internal_functions.R
@@ -647,7 +647,7 @@ is.survival <- function(.name){
 globalVariables(c("L95", "U95", "fit_id", "Total",
 									"OR", "HR", "Coefficient", ".", ".id", "var", "value",
 									":=", "Mean", "SD", "Median", "Q3", "Q1", "IQR", "Formatted", 
-									"w", "Freq", "g", "total_prop", "Prop", "index_total", "vname"))
+									"w", "Freq", "g", "total_prop", "Prop", "index_total", "vname", "Combined"))
 
 
 # Workaround ::: as summary.formula not (yet) exported from Hmisc