+++ NEWS.md
@@ -1,7 +1,7 @@
 # finalfit 0.9.1
 
 * `summary_factorlist()` rework. Now supports any number of factor levels in dependent. #14 #15
-
+* `summary_factorlist()` now provides total value for continuous variable. #17
 
 # finalfit 0.9.0